Expose nicer power user features for rewriting
Not currently used (but useful for debugging). In the long term I don't
think this rewriting automation is really working. It's a pile of hacks
(albeit useful ones) that doesn't work reliably enough, and I'd like to
replace it with something reflective that handles `>>` ("seq", binds
without a dependency) really well.