Add dependencies for iris/stdpp, delete broken files.
iris/stdpp do not follow the 'vendor' scheme, so they've been put in the
external/ subdirectory, and the tweaks to _CoqProject are manually generated.

Also add a bunch of silences to warnings. This is a bit of a hack for now.

Had to tweak various notations and substitute use of omega with lia to resolve
incompatibilities.