Initial port of Iris weakest precondition.
This adapts the Iris definition of weakest precondition to the shallow
embedding used in Argosy. This is otherwise the "standard" Iris weakest
precondition: there is no separate halt condition, unlike CHL. Iris already has
a mechanism for postulating (and enforcing) invariants, which I believe can be
re-used for halt conditions. This commit only adapts the definition and proves
the basic structural rules, the adequacy proof still needs to be adapted.