Finish variable semantics
Simplified things by just separating out the types of variables from the
type of indices; this solves the problem of checking equality of
variables (which are actually just codes, the type index isn't
important). The type mapping is static and the mapping itself is
complete, unlike for other data types, so no checks are needed in the
semantics. The only inconvenience of this approach is that the GADT
defined variables alongside their types, whereas here the definitions
are separate.

In Haskell we'll just have to unsafeCoerce, but the GADT approach
would've had the same problem since Var would have lost its nice type
indices in the extraction process.