Provide an explicit model for ByteString
Rather than defining them axiomatically and then assuming properties, we
simply use a model and then carefully interact with an API that at
extraction time can be replaced by an efficient implementation using
actual Haskell ByteStrings.

The only problem is that the model of numbers as modular arithmetic is
extremely painful to prove properties about.