Fix the semantics of Bind. Prove wp rules for bind, loop, and StatDb spinlock.
It used to be the case that 'Ret v' could take a dummy step, so when stepping a
'Bind e f', we would always first step e, then check if the result was a Ret
and do the substitution if necessary. That's a bit strange and makes it hard to
prove the expected sequencing rule for Bind.

Now, Ret does not take a step, and the rule for Bind cases on whether e is a
Ret or not.