Finish proving internal versions of adequacy theorems.
The new versions allow the 'crash invariant' to be an Iris propositions,
whereas the previous versions required them to be pure Coq propositions.
The difference amounted to doing the inductive reasoning about
the repeated sequences of crashes at the meta-Coq level or 'inside' the
separation logic.

Soundness for the latter was trickier to prove (but will be easier to use)
because the proof involves counting the sum of the steps taken
by the initial program and all of the partial executions of recovery.