New adequacy proof for proving a sequence of programs refines another (modulo one admit).
This version implicitly does some of the inductive reasoning that the framework
previously did using the abstraction relation. The benefit over the previous
version is that we do not have to drop out of the logic to state the
abstraction relation.

Still need to actually instantiate this general theorem to get an eaiser to use
version for ExMach examples to see that this all actually paid off, but the
commit was getting big enough as is.