Apply the framework to the StatDB example to show that the impl gives refinements.
The notion of refinement here is only about final observable behavior (so far),
not trace outputs.

It is very verbose for such a trivial example, but there are some re-usable
patterns that could be extracted out, and part of this is some 'standard
boilerplate' -- the relative costs of which will not seem so bad for more
serious examples (hopefully).

This result still depends on one admit in the generic CSL adequcy theorem.
Namely, one that shows the compiled program has no errors -- but I know how to
prove that.