Finish as much as possible the write ahead log proof (it turns out not be true!)
The result is not true because of our current definition of termination for
concurrent programs, which says that a program terminates when the initial
thread returns a value. The problem arises when a parent thread terminates after
a writer has set the commit bit but before it has propagated writes to the main
entries. Then, if a crash occurs at the beginning of the next program's
execution, recovery will commit the writes from the previous program's
interrupted writer. This does not correspond to an observable behavior of the
new program.

The solution is to allow finishing execution of a program to transform state,
say by adding a `finish_step`, just as we have a `crash_step`, and then have a
default procedure like recovery that always runs upon startup (Tej suggests
perhaps running recovery itself). We already resolved at today's meeting to do
this anyhow.