Add operations to block until all other threads finish. Finish log proof.
We now track the number of running threads with a counter. When a thread is
spawned, this counter is incremented, and we append a command called
"Unregister" to the end of the spawned thread that decrements the counter.  A
command called "Wait" blocks until this counter reaches 1 (meaning all other
threads have called Unregister). Crashing and finishing execution resets this
counter to 1.  A "finish_step" relation has been added to layers but is not yet
run at the end of execution.

Client programs in an abstract layer should not directly call Unregister/Wait.
Rather, when compiling we adjoin a call to "Wait" to the end of the first
thread of the compiled program. This is needed for implementations like the
write ahead log, where correctness requires that all client operations in other
threads finish before launching a new program. (The plan is also to
subsequently let layer implementations have a special 'Close()' proc that would
be called after "Wait"). (In principle clients could be allowed to call 'Wait',
but the current refinement theorem forbids this.)

At the logical level, we now have two assertions called
"Registered" and "AllDone". Launched threads get a "Registered" permission.  We
also have a triple for "Wait" which gives an "AllDone" in the post-condition.
Proofs use a rule that says Registered * AllDone |- False to be able to exploit
the fact that other threads should not be running after a call to Wait.

This is used to complete the write-ahead-log example.