Split mailbox lockstatus into 3 states.
Previously, a mailbox could be either locked or unlocked. A pickup would
implicitly lock a mailbox, and a delete triggered an error if it executed on an
unlocked mailbox. This meant that clients which had unordered calls to pickup
and delete would trigger undefined behavior if the delete happened before
the pickup, as expected.

However, if the pickup by coincidence happened first, this particular execution
would not directly result in undefined behavior. It turns out to be difficult
in the implementation proof to argue that a given scenario is impossible
because some hypothetical alternate order of operations would have triggered
undefined behavior.

Instead, this commit now splits the mailbox status into 3 possible states:
MUnlocked, MLocked, and MPickingUp. The first stage of pickup transitions to
MPickingUp and then finishes by transitioning to MLocked. A delete triggers an
error if the mailbox is either MUnlocked or MPickingUp, so unordered Pickups
and Deletes trigger undefined behavior either way they end up being scheduled.