Add initialization to MailAPI.
Add a boolean flag to Mail state which captures whether the locks have been
initialized yet, along with a new operation called "Open" to set this flag.
When the flag is false, any operation other than Open triggers undefined
behavior.

Finishing program execution unsets the flag, while crashing sets it (because
recovery will have initialized the locks, at least in the current
implementation, however arguably we should consider changing this).

Note: the invariants for the refinement proof need to be updated to reflect
that when this flag is false, the locks/global are not set up properly.