Change how two disk faults.
In initial version, had a separate operation that would try to fail a disk. The
idea being that we could simulate the "background" failure steps that we used
in the Argosy version this way. However, that would let us preclude failures
from happening during recovery.

Now, switch to something like the Argosy version, where there's a pre-step that
tries to fail a disk before performing an operation. However, unlike Argosy, we
will only do this before disk operations (that seems sufficient).