Shorten many slow Qeds for refinement proofs.
A big part of the problem seems to be that the typeclass proofs showing certain
resource algebras were available to use were huge and repeated over and over
again within the proof terms. This was painfully obvious upon looking more
carefully at the terms again.

The solution is to just declare opaque instances for each of these goals.