Define a weakest pre with a crash invariant.
Define "wpc e {{ P }} {{ Q }}" where Q is now a crash invariant
(or, what Argosy calls a halt invariant).

The key is a lemma:

```
Lemma wp_wpc_inv (N: namespace) s E e Φ Φc :
  inv N Φc ∗ WP e @ s ; E {{ Φ }} ⊢ WPC e @ s ; E {{ Φ }} {{ ▷ Φc }}.
```

which shows if you have an Iris invariant and a normal weakest
pre, you can derive a wpc where the Iris invariant is opened and moved into
the crash invariant.

Not sure yet if this is an assertion we will actually use much, but I think it
might help structure some of the soundness results and make the refinement
conditions easier to understand.