Messy implementation of struct loading
The executable part isn't so bad but the typechecking is rather hairy.
It would be nice to either do this in a simpler way or typecheck each
instance separately, rather than giving a generic proof.

Note that append_log uses these new type system features, but the
typechecker doesn't really support them so there's a manual proof (so
the file isn't actually auto-generated). Furthermore, even the type
signatures have to be modified to incorporate the new struct pointer
types.

Having a distinction between pointer references and struct references
purely at the type level makes this much worse than it should be. We
could try more tightly integrating the types into the heap, or at least
building in the notion of flattening. More extreme would be to get rid
of struct values and only have pointers (in the model - Go code could
still use struct values).