Make u64 and u32 instances of Interface.word
This automatically lifts all the functions and theorems to them, so we
can use the word library in a straightforward way as it was intended
(that is, through Interface.word and Properties).

The only caveat is that sometimes we run into unifying word.rep ?word
with u64; Canonical Structures should solve this in theory, but I'm not
sure how to realize that.