Make some progress reasoning about blocks
The disk model is definitely messy and should be fixed to more closely
resemble in-memory values. The only caveat is that we shouldn't change
the underlying state type for disks too drastically (currently it's
reasonable, a map from Z to (vec u8 4096)). Ideally all the changes
would happen in the state interpretation, we do one proof of soundness
against the semantics, and after that we have nicer points-to facts.

I'm thinking that we should start thinking of the disks as being
byte-based, with the 4096-byte blocks being an invariant of the size
(already the lack of holes is an invariant) and with reads and writes
being atomic over 4096 addresses simultaneously.