Add external value types
This makes things slightly messy, but it does seem to work.

We have one serious annoyance, which is that in order to generically
prove that the zero value has the right type, I had to make the zero
value for external values type-check. This breaks the abstraction we
wanted; really, in the model we shouldn't support taking the zero value
of an external type (which is annoying to change now, but it should
work).