Re-write heap lang semantics in an algebraic style
Similar to the Kleene-algebra transitions library, define a type of
transitions with some basic primitives and a general bind combinator,
then define everything as a function from an expression to this
transition denotation.

Transitions are expressed as an inductive to facilitate future exection
and extraction, but also have the obvious relational denotation. There's
currently a lot of duplication between the transitions (which are used
to write things down) and the relations (which are used in the proofs),
which we'll have to fix at some point. They might be unifiable with a
common typeclass.

While these are productive to write, we need to work out how to do
proofs with them. I'm hoping something like monadic/functor
simplification can cut through a lot of the helper definitions and
reduce things down to their essence (eg, you end up with fmaps and binds
of pure expressions all over the place, which can be reduced to a single
expression, or you have cases where there are multiple error checks in a
row that could be collapsed).