Start to define trace refinement.
One issue is that right now the type of traces depends on the external
operations of a layer (since we allow extensions to add in new values, and
trace events contain values). Probably best to change to make traces events
only contain literals, which is what the semantics is supposed to enforce
anyway.