Add wpc specs for circular log
No abstraction relation or proofs yet.

Note that the code and proof assume single-threaded usage of the
circular log, which may not quite be true in the wal (there are a
handful of reads of diskEnd that happen with a lock while appending to
the log; these reads may need to instead access a cache in the wal
protected by a lock instead).