Remove ExtV from Goose
The one complication this introduces is that not every type has a
meaningful zero value, so val_ty (zero_val t) t is no longer always
true. Instead, types must not contain external types to be well-typed.

Go enforces no such restriction, but it has a similar issue where nil
for an opaque type which is a pointer or interface doesn't satisfy its
usual invariants, if you used the API.