Upgrade dependencies
Adapt to https://gitlab.mpi-sws.org/iris/iris/-/merge_requests/358,
which removes a coercion from iProp to Prop (stating that the iProp is
always valid). Instead, we now explicitly write ⊢ P when P is
universally valid. This notation also automatically parses the argument
in bi_scope, which removes the need to use %I for top-level theorems.