Stop using diskEnd_is to link to circ state
Can't put diskEnd_is in an invariant, because it must be fully owned by
the circular buffer during appends; however, we can still refer to the
abstract state of the circular buffer through the ghost variable that
has its whole state.

To connect the lower-bound ghost resources to the state, we export an
accessor for the half-ownership fragments in the circular invariant
itself, which are connected to the state in the invariant through a
ghost variable.