Remove namespaces and "fuel" from crash invariants
Crash invariants do not have a namespace associated anymore. They can be opened
around a wpc no matter what the mask is.

Opening and allocating crash invariants does not decrease the natural number
"fuel" indexing wpc anymore. Instead, that natural number corresponds to a
hierarchy of crash invariant levels, somewhat like universe levels in type
theory. When opening a creating invariant, the wpc's subscript must be *below*
the level of the crash invariant. Conversely, when using a cfupd to remove
a crash obligation from a wpc's crash condition, the cfupd's level must be
*above* the wpc's subscript.