Flush specs for single async inode
Two versions are given: an abstract HOCAP version and a simple consequence. The
latter says that if one has a lower bound on the "total" blocks in the inode
(both durable and buffered) before the flush, then after the flush this is
promoted to a lower bound on the durable blocks.