Prove stable_sound in new generation
This is super annoying becuase it turns out we could have preserved the
entire set of stable transactions at no cost, and it even would have
given us stability of `diskEnd_txn_id` (which we currently prove as a
special case of being the last transaction, which is automatically
`stable_sound`).

`stable_sound` is also defined in a really Coq-unfriendly way. It should
at least be stated as a property of each individual transaction and then
lifted to the set, which would help us since we take a subset and keep
them stable. `nextDiskEnd_inv` similarly has no abstraction around the
read-only map, which also makes it painful to work with.