Keep invG the same after a crash in wpr
Perennial 1 style idempotence reasoning opened up all invariants upon crash and
then introduced a fresh invG instance, which controls the implicit gname used
in fupds/invariants.

This was not exploited anymore in Perennial 2 as of commit cb960af959, which
removed the ability to leave invariants open when proving a crash condition,
since this was no longer needed after the introduction of ncinv/ncfupd.

Despite that, the fresh invG was still being generated after each crash in the
definition of wpr. These fresh invG's are incompatible with the planned use in
Grove, where we want to have multi-node invariants which should remain in
force even if a single node crashes.

This commit removes the generation of fresh invG in wpr as groundwork for
Grove.