Started figuring out how to steal PreCond from old requests via double fupd
The Get() RPC is actually too easy, because we could just change its
spec to say { True } Get() { RET v; (kv [[γ]]↦ some_v -∗ some_v = v) }
and then don't need to ever send the (kv [[γ]]↦ v) over the network.

For Put(), this won't work. For Put(), idempotency holds because a new
Put() will overwrite the old Put(). So, when we run a put, we need to
steal the old Put()'s kv ptsto fact by opening up its request invariant
and taking either its PreCond (if the old Put() never ran) or taking its
PostCond (in case the old Put() already ran). We can only steal the old
requests's (Pre \/ Post) if we know that it'll never be run again; i.e.
we need to update the lastSeqM in order to take the Pre out. This means
that the fupd where we go Pre -> Post now actually also has to invoke
another fupd to even get the Pre.

The the crash obligation is a bit sketchy (it's fupd that gives another
fupd). Knowing that the durable client seqno for Cid is Seq tells us
that as soon as we update the lastSeq on the server to be at least
(Seq - 1), then we can safely steal the old request's (Pre \/ Post).