Start to add typing for pure lists
Lists are represented using the usual encoding in terms of pairs and sums.
However, the typing rules need to be adjusted to preclude doing struct stores
with list types, because their dynamic size means the usual flattening of pairs
does not do the right thing.

These are purely for specifications, they're not intended to be the target
of any real Goose translation.