Use unproven spec for incr service in MakePreparedRequest proof.
Need to weaken spatial context to the loop invariant for the
MakePreparedRequest proof, just like in the proof of MakeRequest.

Had to do a bit of work to turn a
`(readonly (ck ↦[ShortTermIncrClerk.S :: "req"] <struct literal>))`
into a bunch of readonly ptstos for all of the fields in the req struct,
to satisfy the read_request precondition of RemoteProcedureCall_spec.