Progress in incr_proxy_proof; need to prove MakeFreshIncrClerk;
Also need to structure proof of increment_proxy_core so that
`if (unprepared) { prepare } ;; rest` can have just one proof for
the rest portion.

Need to know that the seqno being used for the RPC request can safely be
incremented in order to use make_request; probably can strengthen
ShortTermClerk to be a OneTimeClerk, and know that the only seqno it
uses for a request is 1.