Progress on TransactionCoordinator__doTransfer;
need to add additional witness as post of Prepare that the logical
txn_data matches the parameters that were passed in as input;
this will require the coordinator to actually do the one-shot update of
txn_data. Otherwise, the coordinator could call PrepareIncrease with
different parameters for the same tid, which would be wrong.