Trying to use Löb induction to avoid explicitly writing down pure
functions that describe the changes in state caused by functions.
With the reply table, it's would be kinda annoying to write down the
pure Coq function that applies a sequenced Put request to an arbitrary
state of the GoKV server, but all we really need to know is that when we
run the function *again* in the future, we'll get the same behavior. So,
(by lifting the existential for the new state of the server outside of
the Hoare triple), Löb induction lets you know that the proof you're
doing right now *is* the proof you need in order to know that you can
safely append an encoding of the current operation to the current aof.