Note about missing stuff in Get spec;
Need to know that when the latestEpoch is increased, the value that is
initially set for the new per-epoch value is the value that used to be
in latestEpoch. This will allow the caller to know that the new epoch is
going to start out in sync with the end-client (epoch-unaware) points-to.