Update coqutil
Behavior of out-of-range shifts (that is, shifts higher than the
bitwidth) has changed but is customizable so we restore the old
behavior. Before, an out-of-range shift would effectively shift and then
take the low bits, always producing zero. The new default is to follows
RISC-V behavior and mask the shift itself first so that 1 << 64 is
equivalent to 1 << 0 and produces 1.