Avoid repetition by using EnterNewEpoch_spec_wand to join branches;
Seems like something is missing in the proof for the situation in which
the frontend server has `own_unused_epoch` and the ctr server has
already entered that epoch. Previously, the ctr.Clerk would require
either own_unused or own_val, but not the server only optionally takes
in own_unused. What we need to know now is that if frontend.Server has
`own_unused_epoch e`, then the latest epoch cannot be equal to e.