Started proof of become_leader;
for proof elegance, tried adding an explicit predicate that keeps track
of the important facts about configuration changes; something like the
"old_term_max" but for configs instead of terms. This almost worked,
except for some mutual recursion, which leads to later problems because
these predicates are put inside of an invariant. Might be able to get
around the issue with a few steps.