Add proposal ghost state to simplelog; need different P and Pcrash in aof
to support simplelog. Want to keep resources along with file points-to
that we don't care about after a crash.

Encoding between bytes and state is not injective because the same state
can be represented by different sequences of ops. So, the proposal ghost
state tracks the specific (ops, epoch, sealed) state that the byte data
is supposed to represent.

Relatedly, it seems like the injectivity requirements on op/snapshot
encoding can be removed everywhere.