Proof for updating local ephemeral proposal when backup gets ApplyAsBackup RPC;
this is a bit of a "bogus" proof in the sense that backup servers don't
even use the ephemeral proposal. But, if we encoded this fact in the
invariants, then we would need to break the existing own_primary_ghost
abstraction and would need to give as precondition to ApplyAsBackup
a witness that says the server is not primary in that epoch.

I decided to just do the bogus proof instead (unclear if this is
actually less work).