Fix some of applyBackup proof; got stuck re-establishing all the
new pure facts (about read-only optimization ephemeral state) at the
post-waitFn lock release in ApplyAsBackup. To fix that, I started making
a bigger change that would actually involve changing to ApplyAsBackupRPC
spec to require that the server is not a primary, but I realize now that
shouldn't be needed, so will undo some of that and fix the proof a
different way.