Notes about ephemeral proposal, and possible need for (frozen) pre-ephemeral proposal
Current invariants maintain that "nextRoIndex" is the number of RO
operations at the end of the local accepted log. Alas, this is
technically wrong.

nextRoIndex is the number of RO operations at the end of the local
accepted log *that were proposed in the current epoch*. So, instead of
(length of RO suffix) = nextRoIndex, we have RO suffix = (pre-RO suffix) +
(new RO suffix), and (length of new RO suffix nextRoIndex).
This will complicate some of the list length arithmetic, and might
invalidate some other invariants. They will have to be carefully
weakened to make them inductive again.