Checkpoint progress removing ephemeral proposal; realized that the
proof still needs an ephemeral proposal, so I'll have to add back in
some version of it. This time, I can do it right and only keep an
ephemeral proposal on the primary and avoid the headache of keeping it
up to date on the backups.

The ephemeral proposal is ghost state to address the following issue:
when the primary gets an "ApplyRo" RPC from the client (e.g. to do a
Get(k) in a KV system), it reads the state from some log `ops1` of
operations which might not yet be durable. These ops might have not
actually been "proposed" in the pb_protocol sense. The pb_protocol
proposal is updated when the ops are made durable.

So, when reading the not-yet-durable list `ops1`, we need the server to
promise that if anything is proposed prior to a crash, the thing that's
proposed is exactly what the ApplyRo thread reads in memory. The
ephemeral proposal tracks the ops that the server has promised it's
going to try really propose in this crash-generation. In terms of this
ephemeral proposal, one can make everything work out (this is how the
heartbeat-based optimization works).

This huge proof complicatoin can be traced to one aspect of the
implementation:
  a server might be the leader in a particular epoch and put some ops in
  its log in memory, crash, restart, get a duplicate RPC telling it to
  become leader in that same epoch, and then put DIFFERENT ops at the
  same place in the in-memory log. This forces the implementation to not
  send RPCs to backups until AFTER the primary makes its state durable.

Maybe it's time to stop doing this, and reduce the normal-case latency
by 1 disk write in exchange for disallowing a server to restart and
continue as primary in the same epoch.

Somehow making this change to the Go code would greatly simplify the
proof and improve performance a bit. I think we know how we might
approach the more complicated proof challenge, having done most of it in
the case of heartbeats using this ephemeral proposal stuff. At this
point, either we should plan to talk about this ephemeral proposal proof
challenge, or tweak the implementation to do away with it and focus on
what we actually care about in the proof.