Use lease to show that a replica with lease has up-to-date state;
need to clean up the interface of some libraries in terms of their
masks. The proof for this (implicitly) opens 4 invariants at once from
different libraries (config service invariant, core primary/backup
invariant, a lease invariant, and append-only file invariant). It's the
append-only flie invariant that's really posing the problem, because the
API for accessP in pb_definitions is too conservative and assumes that
all invariants are invalidated at this step except for its own. In
reality, only the AOF invariant is invalidated.