Maintain that a replica server really is in the config for its current epoch;
this was unnecessary before i.e. one could run extra backup servers
that aren't in the config. Now, this is not allowed because those
backup servers might be serving read-only ops and so we must be certain
they are getting all the ops from the primary before acking the client.