Improve word's rewriting
The rewrite in [word] that simplified through (uint.Z (word.add ...))
(and similar rewrites for other operations) didn't have a `repeat`,
which meant that it would (sometimes) not fully rewrite through multiple
operations. This is now fixed, and proofs have been updated to deal with
the more powerful tactic.

Where this caused problems was when a manually written lemma was used
after a `word_cleanup`, relying on the particular set of rewrites that
were performed before. These were mostly in old code, and I expect new
code wants a `word` tactic that is as powerful as possible.