Rename Mutex and Cond methods for consistency
Previously, the names acquire/release and acquire_spec/release_spec came
from HeapLang, and mutexes were the only place this convention was used.

For consistency with how we typically name functions and specs (and to
match Go), this commit renames to `Mutex__Lock` and `wp_Mutex__Lock`,
and similarly for Unlock. At the same time we make the Cond methods more
consistent with this convention, although its specs already used the
`wp_` naming style.