Add model and proof of some of sync/atomic
This is only the basic atomic load/store of integers. There's a safe API
implemented using generics for pointers, which can be supported if we
also translate the atomic.Pointer type correctly; the model is no
different, it's only an issue of translation and handling that type as
an alias for a GooseLang pointer.