Sync tulip proof
Note that right now the "debug/moderation" procedures (i.e., dumping states and
forcing re-election to control the leader location) is invoked through an
out-of-band interface (i.e., not through the verified client). This means the
proof does not enforce their correctness. The plan is to add an verified
moderator to make operations outside the transaction protocol. All the
moderation methods (on the moderator and the server sides) should have trivial
specification, but proving against them would still be useful by ensuring no UBs
appear in the implementation.