Make package is_defined and is_initialized systematic (#191)
This PR adds new typeclasses `PkgIsInitialized pkg_name` and `PkgIsDefined pkg_nam` that associate a package with its "initialization" and "definenedness" predicate, respectively. The latter is already automatically generated, while the former is defined in manual proofs but in the absence of global variables or initialization functions consists of `is_defined` for that package and is_initialized for all imported packages. The instances of these typeclasses essentially register a predicate to a package name, accessed via `is_pkg_init` and `is_pkg_defined`.

This PR also adds a new `PkgInfo` typeclass, which is used to get the information that constructs `is_defined` and the list of imports.

This PR depends on goose-lang/goose#65 to create new instances and definitions.

These typeclasses permit some useful automation of initialization predicates, most of which is included here:

- The tactic `basic_pkg_init` defines an initialization predicate.
- `wp_apply` automatically proves any pkg_init preconditions it finds from any found in the persistent context.
- A new `wp_start` tactic automatically introduces the initialization predicate and names only the remaining hypotheses according to the user pattern.

---------

Co-authored-by: Upamanyu Sharma <upamanyu@mit.edu>