Infer own_map arguments from m's type
There is a subtle issue here where Coq infers left to right, so kt and
vt must be inferred before the type of m is known. This makes own_map
harder to use. In practice, supplying kt solves the problem, but it
should not be necessary since we can go from the type of m to a value
for K to finally a type for kt using IntoValTyped.